requirements.txt
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
mysqlclient==2.1.0
Pillow==9.0.0